WIP refactor the steps
