+++ static/js/src/advantage/subscribe/react/components/ModalParts/ModalFooter.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { Row, ActionButton } from "@canonical/react-components";
+
+type ModalFooter = {
+  children: React.ReactChild;
+};
+
+const ModalFooter = ({ children }: ModalFooter) => {
+  return (
+    <footer className="p-modal__footer">
+      <Row className="u-no-padding">
+        <ActionButton
+          className="js-cancel-modal col-small-2 col-medium-2 col-start-medium-3 col-start-large-7 col-3 u-no-margin"
+          appearance="neutral"
+          aria-controls="purchase-modal"
+          style={{ textAlign: "center" }}
+        >
+          Cancel
+        </ActionButton>
+        {children}
+      </Row>
+    </footer>
+  );
+};
+
+export default ModalFooter;
